+++ Backend/src/main/java/com/pz/ankietBud/model/subQuestion/Rating.java
@@ -15,7 +15,7 @@ public class Rating extends Question {
     @Enumerated(EnumType.STRING)
     private Type type;
 
-    public Rating(Integer id, String content, ArrayList<Integer> votes, Integer sum, Integer value_min, Integer value_max, Type type) {
+    public Rating(Long id, String content, ArrayList<Integer> votes, Integer sum, Integer value_min, Integer value_max, Type type) {
         super(id, content, votes, sum);
         this.value_min = value_min;
         this.value_max = value_max;
